requirements.txt
Flask==2.2.5
python-telegram-bot==20.4
gunicorn==21.2.0
